+++ Sources/Kommunicate/Classes/KMConversationViewConfiguration.swift
@@ -20,6 +20,12 @@ public struct KMConversationViewConfiguration {
     public var restrictMessageTypingWithBots = false
     public var toolbarSubtitleText = ""
     public var toolbarSubtitleRating: Float = -1.0
-
+    /// Customize Text color of FAQ button on conversation, conversation list screen
+    public var faqTextColor : UIColor = UIColor.white
+    /// Customize background color of FAQ button on conversation, conversation list screen
+    public var faqBackgroundColor : UIColor = UIColor(hexString: "ffffff", alpha: 0.24)
+    /// Customize the Start New Conversation Button (Background Color, TextColor ) on Conversation List Screen
+    public var startNewConversationButtonBackgroundColor: UIColor? = nil
+    public var startNewConversationButtonTextColor: UIColor = UIColor.white
     public init() {}
 }