+++ Sources/Kommunicate/Classes/Kommunicate.swift
@@ -66,7 +66,9 @@ open class Kommunicate: NSObject, Localizable {
         config.isTapOnNavigationBarEnabled = false
         config.isProfileTapActionEnabled = false
         var navigationItemsForConversationList = [ALKNavigationItem]()
-        let faqItem = ALKNavigationItem(identifier: faqIdentifier, text: NSLocalizedString("FaqTitle", value: "FAQ", comment: ""))
+        var faqItem = ALKNavigationItem(identifier: faqIdentifier, text: NSLocalizedString("FaqTitle", value: "FAQ", comment: ""))
+        faqItem.faqTextColor = kmConversationViewConfiguration.faqTextColor
+        faqItem.faqBackgroundColor = kmConversationViewConfiguration.faqBackgroundColor
         navigationItemsForConversationList.append(faqItem)
         var navigationItemsForConversationView = [ALKNavigationItem]()
         navigationItemsForConversationView.append(faqItem)
@@ -1121,6 +1123,23 @@ open class Kommunicate: NSObject, Localizable {
         }
         (topVc as! KMConversationViewController).updateChatbarText(text: text)
     }
+      
+    /*
+     This method will show/hide assignee's online, offline status and away message on conversation screen when its on top.
+     - Parameters:
+     - hide: boolean to show/hide the views
+     */
+    open class func hideAssigneeStatus(_ hide: Bool) {
+        let pushAssist = ALPushAssist()
+        guard let topVc = pushAssist.topViewController,
+              topVc is KMConversationViewController
+        else {
+            print("Failed to hide assignee status")
+            return
+        }
+        (topVc as! KMConversationViewController).hideAssigneeStatus(hide)
+    }
+    
     // MARK: - Deprecated methods
     
     /**
@@ -1171,6 +1190,7 @@ open class Kommunicate: NSObject, Localizable {
     open class func createSettings(settings: String) -> Bool {
         return KMConfigurationSetter.createCustomSetting(settings: settings)
     }
+    
     
     /**
      Creates a new conversation with the details passed.